# Conflicts:
#	app/imports/ui/pages/landing-page.html
#	app/imports/ui/stylesheets/style.css

+++ app/imports/api/role/Role.js
@@ -35,6 +35,8 @@ export function assertRole(role) {
 // Initialize Roles to ROLENAMES by deleting all existing roles, then defining just those in ROLENAMES.
 
 if (Meteor.isServer) {
-  Roles.getAllRoles().fetch().map(role => Roles.deleteRole(role.name));
-  _.values(ROLE).map(role => Roles.createRole(role));
+  if (Roles.getAllRoles().count() !== 4) {
+    Roles.getAllRoles().fetch().map(role => Roles.deleteRole(role.name));
+    _.values(ROLE).map(role => Roles.createRole(role));
+  }
 }
